dir_with_timestamp

dir_with_timestamp is a wrapper around base::dir(), with some additional capabilities like filtering out folders with identical names but old timestamps.

Usage

dir_with_timestamp(
  path = ".",
  pattern = "",
  full.names = FALSE,
  name_timestamp_seperator = "_",
  timestamp_format = "%Y-%m-%d_%H.%M.%S",
  remove_timestamp = FALSE,
  only_most_recent = FALSE,
  ...
)

Arguments

path

A string with the path to the folder which is to be listed.

pattern

A string with a regex pattern with which to filter.

full.names

Boolean, if true return full paths.

name_timestamp_seperator

A string with the name-timestamp seperator.

timestamp_format

A string with the timestamp format, as input for base::as.POSIXct().

remove_timestamp

Boolean, if true remove timestamp.

only_most_recent

Boolean, if true remove folders with identical name but older timestamps.

...

Parameters that are passed on to base::dir().

Value

A vector of strings with folder paths.