Summarise a model with statistics indicators

Summarise a model with statistics indicators

model_summarise(
  data,
  x = "x",
  y = "y",
  digits = NULL,
  direction = c("wide", "long")
)

Arguments

data

A data frame to summarise

x

x variable name

y

y variable name

digits

integer indicating the number of decimal places (round) or significant digits (signif) to be used.

direction

The wide (default) or long format for the output

Value

A data frame with statistics indicators in columns including:

n: Number of rows

r: Coefficient of correlation

r2: Squared coefficient of correlation

bias: Average amount by which actual is greater than predicted

mse: Average squared difference

rmse: Root mean squared error

rrmse: Relative root mean squared error

Examples

library(dplyr)
#> 
#> Attaching package: ‘dplyr’
#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag
#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union
data <- data.frame(x = 1:10, y = 1:10 + runif(10))
data %>% model_summarise()
#>    n         r       r2       bias       mse      rmse      rrmse
#> 1 10 0.9955813 0.991182 -0.4012284 0.2459003 0.4958834 0.02851129
data %>% model_summarise(digits = 2)
#>    n r   r2 bias  mse rmse rrmse
#> 1 10 1 0.99 -0.4 0.25  0.5  0.03
# Export as long format
data %>% model_summarise(digits = 2, direction = "long")
#> # A tibble: 7 × 2
#>   indicator value
#>   <chr>     <dbl>
#> 1 n         10   
#> 2 r          1   
#> 3 r2         0.99
#> 4 bias      -0.4 
#> 5 mse        0.25
#> 6 rmse       0.5 
#> 7 rrmse      0.03