read_items_from_mifs

read_items_from_mifs reads specific items from the remind mif files.

Usage

read_items_from_mifs(
  mif_filepaths,
  regexs_of_items,
  col_types = "cccccddddddddddddddddddd-"
)

Arguments

mif_filepaths

A vector of strings containing the file paths.

regexs_of_items

A vector of regex patterns identifying the variables to extract.

col_types

A string with the col_type specification for readr::read_delim

Value

A tibble